♻️ refactor: highlightKeyword, sliceTitle 함수 utils로 분리
