+++ frontend/src/pages/SearchList/components/Paper.tsx
@@ -1,7 +1,6 @@
 import styled from 'styled-components';
-import { MAX_TITLE_LENGTH } from '../../../constants/main';
 import { Ellipsis } from '../../../style/styleUtils';
-import { removeTag } from '../../../utils/format';
+import { highlightKeyword, removeTag, sliceTitle } from '../../../utils/format';
 import { IPaper } from '../SearchList';
 
 interface PaperProps {
@@ -14,35 +13,14 @@ const Paper = ({ data, keyword }: PaperProps) => {
 
   const year = new Date(publishedAt).getFullYear();
 
-  // keyword 강조
-  const highlightKeyword = (text: string) => {
-    const rawKeywordList = keyword.trim().toLowerCase().split(/\s/gi);
-
-    return rawKeywordList.length > 0 && rawKeywordList.some((rawKeyword) => text.toLowerCase().includes(rawKeyword))
-      ? text
-          .split(new RegExp(`(${rawKeywordList.join('|')})`, 'gi'))
-          .map((part, i) =>
-            rawKeywordList.some((keywordPart) => part.trim().toLowerCase() === keywordPart) ? (
-              <Emphasize key={i}>{part}</Emphasize>
-            ) : (
-              part
-            ),
-          )
-      : text;
-  };
-
-  const sliceTitle = (title: string) => {
-    return title.length > MAX_TITLE_LENGTH ? `${title.slice(0, MAX_TITLE_LENGTH)}...` : title;
-  };
-
   return (
     <Container>
-      {title && <Title>{highlightKeyword(sliceTitle(removeTag(title)))}</Title>}
+      {title && <Title>{highlightKeyword(sliceTitle(removeTag(title)), keyword)}</Title>}
       {authors && (
         <Content>
           <div>
             <Bold>{authors.length > 1 ? 'Authors ' : 'Author '}</Bold>
-            <Ellipsis>{highlightKeyword(authors?.join(', '))}</Ellipsis>
+            <Ellipsis>{highlightKeyword(authors?.join(', '), keyword)}</Ellipsis>
           </div>
         </Content>
       )}
@@ -85,11 +63,6 @@ const Content = styled.div`
 `;
 
 const Bold = styled.b`
-  font-weight: 700;
-`;
-
-const Emphasize = styled.span`
-  color: #3244ff;
   font-weight: 700;
 `;
