--- frontend/src/utils/format.ts
@@ -1,3 +1,0 @@
-export const removeTag = (text: string) => {
-  return text.replace(/<[^>]*>?/g, ' ');
-};
